% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint.R
\name{CreateSPRINTObject}
\alias{CreateSPRINTObject}
\title{Create the SPRINT object with filtering step}
\usage{
CreateSPRINTObject(counts, location, project = "SPRINT",
  prectage = 0.1, min_total_counts = 10)
}
\arguments{
\item{counts}{Gene expression count matrix (data.frame), p x n -- p is the number of genes and n is the number of cells}

\item{location}{Cell location matrix (data.frame) or two-component of t-SNE/UMAP, n x 2}

\item{project}{Project names}

\item{prectage}{The prectage of cells that are expressed for analysis}

\item{min_total_counts}{The minimum counts for each cell for filtering}
}
\value{
Returns SPRINT object with filtered gene expression matrix
}
\description{
Create the SPRINT object with filtering step
}
