\name{pcor}
\alias{pcor}
\title{Parallel Correlation}
\description{
Sister function to "cor" implementing a parallel Pearson correlation using a replicated data strategy.
}
\usage{
pearsonCoeffs = pcor(data)
}
\arguments{
    \item{data}{double precision 2D array of data}
}
\seealso{
    \code{\link{cor}}
}
\keyword{utilities}
\keyword{interface}