\name{phamming.distance}
\alias{phamming.distance}
\title{Parallel hamming distance}
\description{Calculates the hamming distance matrix between a set of DNA
  sequences provided as a ShortReadQ or DNAStringSet objects.
}
\usage{
  phamming.distance(data, output_filename)
}
\arguments{
  \item{data}{ShortReadQ or DNAStringSet objects}
  \item{output_filename}{results will be stored here as binary data}
}
\details{
  This is an early but fully working prototype.
}
\author{
  University of Edinburgh SPRINT Team
  \email{sprint@ed.ac.uk}
  \url{www.r-sprint.org}
}
\keyword{utilities}
\keyword{interface}

