\name{phello}
\alias{phello}
\title{SPRINT Hello World}
\description{
Simple example function demonstrating adding a method to the SPRINT library. 
Prints a 'hello from processor n' message.
}
\usage{
phello()
}
\arguments{
None
}
\seealso{
\code{\link{SPRINT}}
}
\author{
University of Edinburgh SPRINT Team
    \email{sprint@ed.ac.uk}
    \url{www.r-sprint.org}
    }
\keyword{utilities}
\keyword{interface}