\name{pterminate}
\alias{pterminate}
\title{SPRINT close}
\description{
Shutdown SPRINT and MPI. Must be used with every script that invokes the 
SPRINT package.
}
\usage{
pterminate()
}
\arguments{
None
}
\seealso{
\code{\link{sprint-package}}
}
\author{
University of Edinburgh SPRINT Team
    \email{sprint@ed.ac.uk}
    \url{www.r-sprint.org}
    }  
\keyword{utilities}
\keyword{interface}
