\name{ptest}
\alias{ptest}
\title{Parallel "Hello World"}
\description{
Simple "Hello World" function to test the correct installation of the SPRINT
package on multiple processors.
}
\usage{
ptest()
}
\arguments{
None
}
\seealso{
\code{\link{sprint-package}}
}
\author{
University of Edinburgh SPRINT Team
    \email{sprint@ed.ac.uk}
    \url{www.r-sprint.org}
    } 
\keyword{utilities}
\keyword{interface}
