% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcestimating.R
\name{sprint.fit}
\alias{sprint.fit}
\title{Fitting the count-based spatial model to estimate the parameters}
\usage{
sprint.fit(model0, maxiter = 100, tol = 1e-05, verbose = FALSE)
}
\arguments{
\item{model0}{The initial model fitted by glm function}

\item{maxiter}{Iteration}

\item{tol}{Tolerance}

\item{verbose}{Output fitting information}
}
\description{
Fitting the count-based spatial model to estimate the parameters
}
