% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctesting.R
\name{sprint.test}
\alias{sprint.test}
\title{Testing multiple kernel matrices}
\usage{
sprint.test(object, check_positive = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{SPRINT object}

\item{check_positive}{Check the kernel matrix is positive or not}

\item{verbose}{Output fitting information}
}
\description{
Testing multiple kernel matrices
}
