\name{pterminate}
\alias{pterminate}
\title{SPRINT close}
\description{
Function which indicates the end of the use of the SPRINT library. It 
terminates the use of MPI and shut down the SPRINT library. It must be used 
with every script that invokes the SPRINT package.
}
\usage{
pterminate()
}
\arguments{
None
}
\seealso{
\code{\link{SPRINT}}
}
\author{
University of Edinburgh SPRINT Team
    \email{sprint@ed.ac.uk}
    \url{www.r-sprint.org}
}
\keyword{utilities}
\keyword{interface}