\name{ptest}
\alias{ptest}
\title{SPRINT Installation Test}
\description{
Function that test the correct installation of the SPRINT library. It simply 
prints a message identifying each processor in the compute cluster.
}
\usage{
ptest()
}
\arguments{
None
}
\seealso{
\code{\link{SPRINT}}
}
\author{
University of Edinburgh SPRINT Team
    \email{sprint@ed.ac.uk}
    \url{www.r-sprint.org}
    }
\keyword{utilities}
\keyword{interface}